% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{compress.matrix}
\alias{compress.matrix}
\title{Compress a matrix using the truncated SVD}
\usage{
compress.matrix(mtx, rank = NA, ratio = NA, avg.error = NA)
}
\arguments{
\item{mtx}{matrix}

\item{rank}{rank of the truncated SVD}

\item{ratio}{compression ratio}

\item{avg.error}{average error}
}
\value{
compressed.matrix object with decomposition, rank, error, avg.error
}
\description{
Compress a matrix using the truncated SVD. Exactly one of rank, ratio and
avg.error should be specified. The compression ratio will be at least the
given value, or the average error will be at most the given value. The
average error is the Frobenius norm of the error divided by the square root
of the number of elements in the matrix.
}
\examples{
compress.matrix(matrix(1:12, 4, 3), rank=2)
compress.matrix(matrix(1:12, 4, 3), ratio=1.1)
compress.matrix(matrix(1:12, 4, 3), avg.error=0.5)
}
