% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{decompress.matrix}
\alias{decompress.matrix}
\title{Decompress a compressed.matrix object}
\usage{
decompress.matrix(cmtx)
}
\arguments{
\item{cmtx}{compressed matrix}
}
\value{
uncompressed matrix
}
\description{
Decompress a compressed.matrix object
}
\examples{
decompress.matrix(compress.matrix(matrix(1:12, 4, 3), rank=2))
}
