% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg-manipulation.R
\name{fix.image}
\alias{fix.image}
\title{Fix image}
\usage{
fix.image(mtx)
}
\arguments{
\item{mtx}{numeric matrix}
}
\value{
thresholded matrix
}
\description{
Thresholds values above by 1 and below by 0.
}
\examples{
fix.image(matrix(seq(-0.5,1.4,0.1), 5, 4))
}
