% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{get.frobenius.norms}
\alias{get.frobenius.norms}
\title{Compute Frobenius norms for with low-rank approximation errors}
\usage{
get.frobenius.norms(sigmas)
}
\arguments{
\item{sigmas}{singular values}
}
\value{
list of ranks (k) and norms (norm)
}
\description{
Associated with each truncated SVD is an approximation error. This function
computes the approximation error for each rank.
}
\examples{
get.frobenius.norms(10:1)
}
