% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{get.singular.values}
\alias{get.singular.values}
\title{Compute singular values}
\usage{
get.singular.values(mtx)
}
\arguments{
\item{mtx}{matrix}
}
\value{
numeric vector of singular values in decreasing order
}
\description{
Compute singular values
}
\examples{
get.singular.values(matrix(1:12, 4, 3))
}
