% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg-manipulation.R
\name{grayscale}
\alias{grayscale}
\title{Convert a colour image to grayscale}
\usage{
grayscale(img)
}
\arguments{
\item{img}{colour image as a 3D array}
}
\value{
matrix corresponding to grayscale image
}
\description{
Convert a colour image to grayscale
}
\examples{
grayscale(jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg")))
}
