% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{is.compressed.matrix}
\alias{is.compressed.matrix}
\title{Check if an object is a compressed matrix}
\usage{
is.compressed.matrix(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE <=> object is a compressed matrix
}
\description{
Check if an object is a compressed matrix
}
\examples{
is.compressed.matrix(compress.matrix(matrix(1:12, 4, 3), rank=2))
is.compressed.matrix(42)
}
