% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg-manipulation.R
\name{load.jpeg}
\alias{load.jpeg}
\title{Load a JPEG}
\usage{
load.jpeg(filename)
}
\arguments{
\item{filename}{JPEG file name}
}
\value{
3D array corresponding to image
}
\description{
Load a JPEG
}
\examples{
load.jpeg(system.file("img", "Rlogo.jpg", package="jpeg"))
}
