% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-svd.R
\name{summary.compressed.matrix}
\alias{summary.compressed.matrix}
\title{Summarize compressed matrix}
\usage{
\method{summary}{compressed.matrix}(object, ...)
}
\arguments{
\item{object}{compressed matrix}

\item{...}{other arguments}
}
\description{
Prints a summary of the compressed matrix
}
\examples{
summary(compress.matrix(matrix(1:12, 4, 3), rank=2))
}
