% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg-manipulation.R
\name{view.image}
\alias{view.image}
\title{Plot image}
\usage{
view.image(img)
}
\arguments{
\item{img}{image to plot}
}
\description{
Plot image
}
\examples{
view.image(load.jpeg(system.file("img", "Rlogo.jpg", package="jpeg")))
view.image(grayscale(load.jpeg(system.file("img", "Rlogo.jpg", package="jpeg"))))
}
